feat: add VM for IntroHistoryView
